+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Currently supported languages:
     - English
     - Brazilian Portuguese
+- Team indicator in 2v2 minigames
 
 ### Improved / Changed
 - The main menu can now be navigated with keyboard/controller